Remove Babel requirement.
Node 4+ supports many ES6 features, the only one we were using which
required Babel was `import` and that is unnecessary.

Using Babel here adds complexity to the client project for no gain, so
let's remove it and simplify things.

As this is an internal developer tool we are more than OK to require
Node 4 or even 5+ for any features we want to use.
