+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "baseplate-core",
-  "version": "1.0.1",
+  "version": "1.1.0-pre",
   "description": "Baseplate core. A tool for helping build prototypes and pattern libraries.",
   "author": "Made Media Ltd. <developers@mademedia.co.uk> (made.media)",
   "contrbutors": [
@@ -30,9 +30,6 @@
     "lint"
   ],
   "dependencies": {
-    "babel-core": "6.4.5",
-    "babel-preset-es2015": "6.3.13",
-    "babel-register": "6.4.3",
     "color": "0.11.1",
     "compression": "1.6.1",
     "errorhandler": "1.4.3",
@@ -51,7 +48,6 @@
   "license": "MIT",
   "devDependencies": {
     "ava": "0.11.0",
-    "babel-polyfill": "6.3.14",
     "clean-css": "3.4.9",
     "nsp": "2.2.0",
     "pre-commit": "1.1.2",
