+++ server/helpers/index.js
@@ -1,2 +1,2 @@
-import requireDirectory from 'require-directory';
+var requireDirectory = require('require-directory');
 module.exports = requireDirectory(module);
