+++ server/helpers/loremWords.js
@@ -1,8 +1,8 @@
-import random from 'lodash/random';
-import capitalize from 'lodash/capitalize';
-import loremIpsum from 'lorem-ipsum';
+var random = require('lodash/random');
+var capitalize = require('lodash/capitalize');
+var loremIpsum = require('lorem-ipsum');
 
-module.exports = (min, max) => {
+module.exports = function (min, max) {
     return capitalize(loremIpsum({
         count: random(min, max),
         units: 'words'
