+++ server/helpers/repeatRange.js
@@ -1,4 +1,4 @@
-import times from 'lodash/times';
-import random from 'lodash/random';
+var times = require('lodash/times');
+var random = require('lodash/random');
 
 module.exports = (min, max, block) => times(random(min, max), i => block.fn(i)).join(' ');
