+++ server/routes/index.js
@@ -1,2 +1,4 @@
-import requireDirectory from 'require-directory';
+'use strict';
+
+var requireDirectory = require('require-directory');
 module.exports = requireDirectory(module);
